Added TransitionFunction and tests.
TransitionFunction encapsulates a transition function for a CA. It is
its own class because it will need to support sexual reproduction and
keeping track of its fitness (for sorting).

StateProvider interface is the interface that a CA must implement; it
was extracted from CA proper to decouple TransitionFunction from CA
directly.